// frontend-vue/src/client/types.gen.ts
export type Pagination = {
    /**
     * Page
     */
    page: number;
    /**
     * Perpage
     */
    perPage: number;
    /**
     * Total
     */
    total: number;
};